Refactor: todoModal 버튼 컴포넌트 삭제
